drastic fix

+++ packages/351elec/sources/scripts/postupdate.sh
@@ -100,9 +100,14 @@ if [ -x /storage/.config/packages/drastic/uninstall.sh ]; then
         /usr/bin/351elec-es-packages remove drastic
 fi
 # forced drastic removal to allow upgrade after every image update
-rm -f /storage/roms/gamedata/drastic/drastic.sh
-rm -f /storage/roms/gamedata/drastic/aarch64/drastic/drastic
-ln -sf /storage/roms/gamedata/drastic /storage/drastic
+if [ ! -d "/storage/roms/gamedata/drastic" ]
+then
+  mkdir "/storage/roms/gamedata/drastic"
+  ln -sf /storage/roms/gamedata/drastic /storage/drastic
+else
+  rm -f /storage/roms/gamedata/drastic/drastic.sh
+  rm -f /storage/roms/gamedata/drastic/aarch64/drastic/drastic
+fi
 
 ## 2021-07-02 (konsumschaf)
 ## Change the settings for global.retroachievements.leaderboards
